Move files into correct folder.
